Moved internal header files to public include directory

+++ include/shaga/internal_halfsiphash.h
@@ -40,13 +40,13 @@ All rights reserved.
 	val = static_cast<uint32_t> (HALFSIPHASH_DATA_SIZE & 0xff) << 24;	\
 	switch (HALFSIPHASH_DATA_SIZE & 3) {								\
 		case 3:															\
-			val |= static_cast<uint64_t> (in[2]) << 16;					\
+			val |= static_cast<uint32_t> (in[2]) << 16;					\
 			SHAGA_FALLTHROUGH;                                     		\
 		case 2:															\
-			val |= static_cast<uint64_t> (in[1]) << 8;					\
+			val |= static_cast<uint32_t> (in[1]) << 8;					\
 			SHAGA_FALLTHROUGH;                                    		\
 		case 1:															\
-			val |= static_cast<uint64_t> (in[0]);						\
+			val |= static_cast<uint32_t> (in[0]);						\
 			SHAGA_FALLTHROUGH;                                      	\
 		case 0:															\
 			break;														\
